Refactor export components and remove unused files
- Deleted the ExportComponent and its associated files, including the customizer and utility functions, to streamline the codebase and improve maintainability.
- Updated import paths in various components to reflect the removal of the ExportComponent, ensuring continued functionality.
- Made minor adjustments to the logs and metrics components for consistency in export functionality.

These changes enhance the overall structure of the application by removing redundant code and improving the clarity of component dependencies.
