+++ src/app/components/common/export/customizer/customizer.component.scss
