+++ src/app/components/common/export/export.component.scss
