+++ src/app/components/common/export/utils/field-analysis.utils.ts
