+++ src/app/components/common/table/table.component.html
@@ -67,7 +67,7 @@
             </th>
             <td mat-cell *matCellDef="let row">
                 <ng-container *ngIf="isGroupRow(row)">
-                    <div class="group-row-content muted italic" *ngIf="column.name === (groupDisplayColumn || columns[0].name)">
+                    <div class="group-row-content" *ngIf="column.name === (groupDisplayColumn || columns[0].name)">
                         {{row.name}} ({{row.count}})
                     </div>
                 </ng-container>
